documentation edits

+++ Source/Meadow.Foundation/Controllers/PID/IdealPidController.cs
@@ -15,7 +15,7 @@ namespace Meadow.Foundation.Controllers.PID
             var dt = now - _lastUpdateTime;
             // seconds is better than ticks to bring our calculations into perspective
             //var seconds = (float)(dt.Ticks / 10000 / 1000);
-            var seconds = ((float)dt.Ticks / 10000f / 1000f);
+            var seconds = dt.Ticks / 10000f / 1000f;
 
             // if no time has passed, don't make any changes.
             if (dt.Ticks <= 0.0) return _lastControlOutputValue;
